Split root project in three modules: core, wiro, slick
