+++ backend/core/src/main/scala/io.buildo.toctoc/core/authentication/AuthenticationError.scala
@@ -1,0 +1,9 @@
+package io.buildo.toctoc
+package core
+package authentication
+
+import io.buildo.enumero.annotations.enum
+
+@enum trait AuthenticationError {
+  object InvalidCredential
+}
