+++ backend/core/src/main/scala/io.buildo.toctoc/core/authentication/Credential.scala
@@ -1,0 +1,5 @@
+package io.buildo.toctoc
+package core
+package authentication
+
+trait Credential
