+++ backend/core/src/main/scala/io.buildo.toctoc/core/authentication/Subject.scala
@@ -1,0 +1,7 @@
+package io.buildo.toctoc
+package core
+package authentication
+
+trait Subject {
+  def ref: String
+}
