+++ backend/core/src/main/scala/io.buildo.toctoc/core/authentication/token/Token.scala
@@ -1,0 +1,11 @@
+package io.buildo.toctoc
+package core
+package authentication
+package token
+
+import java.time.Instant
+
+trait Token[T] extends Credential {
+  def value: T
+  def expiresAt: Instant
+}
