--- backend/src/main/scala/io/buildo/toctoc/authentication/AuthenticationError.scala
@@ -1,7 +1,0 @@
-package  io.buildo.toctoc.authentication
-
-import io.buildo.enumero.annotations.enum
-
-@enum trait AuthenticationError {
-  object InvalidCredential
-}
