--- backend/src/main/scala/io/buildo/toctoc/authentication/Credential.scala
@@ -1,3 +1,0 @@
-package  io.buildo.toctoc.authentication
-
-trait Credential
