--- backend/src/main/scala/io/buildo/toctoc/authentication/Subject.scala
@@ -1,5 +1,0 @@
-package io.buildo.toctoc.authentication
-
-trait Subject {
-  def ref: String
-}
