--- backend/src/main/scala/io/buildo/toctoc/authentication/token/Token.scala
@@ -1,10 +1,0 @@
-package io.buildo.toctoc.authentication.token
-
-import java.time.Instant
-
-import io.buildo.toctoc.authentication.Credential
-
-trait Token[T] extends Credential {
-  def value: T
-  def expiresAt: Instant
-}
